Search bar integrated and working in header, top-bar renamed to header.
