+++ src/app/components/shared/header/header.component.html
@@ -56,12 +56,6 @@
         </div>
         <div class="search-container">
             <app-search-bar [searchString]="_query"></app-search-bar>
-            <!--
-            <mat-form-field>
-                <mat-label>Input</mat-label>
-                <input matInput [value]="_query">
-            </mat-form-field>
-            -->
         </div>
 
         <div class="pill-bar-container">
