+++ src/app/components/shared/header/header.component.scss
