+++ src/app/components/shared/header/header.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { TopBarComponent } from './top-bar.component';
+import { TopBarComponent } from './header.component';
 
 describe('TopBarComponent', () => {
   let component: TopBarComponent;
