+++ src/app/components/shared/search-bar/search-bar.component.ts
@@ -1,8 +1,14 @@
-import {Component, ElementRef, Input, OnChanges, SimpleChanges, ViewChild} from '@angular/core';
+import {Component, ElementRef, Input, OnChanges, SimpleChanges, ViewChild, PLATFORM_ID, Inject} from '@angular/core';
 import {FormControl, FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {MatAutocompleteModule} from '@angular/material/autocomplete';
 import {MatInputModule} from '@angular/material/input';
 import {MatFormFieldModule} from '@angular/material/form-field';
+
+import { isPlatformBrowser } from '@angular/common';
+
+import { AfterViewInit, HostListener } from '@angular/core';
+
+import { MatAutocompleteTrigger } from '@angular/material/autocomplete';
 
 @Component({
   selector: 'app-search-bar',
@@ -16,7 +22,7 @@ import {MatFormFieldModule} from '@angular/material/form-field';
   templateUrl: './search-bar.component.html',
   styleUrl: './search-bar.component.scss'
 })
-export class SearchBarComponent implements OnChanges{
+export class SearchBarComponent implements OnChanges,  AfterViewInit{
   @Input() searchString = new FormControl('');
   isPanelOpened = false;
   searchFocus = false;
@@ -30,14 +36,39 @@ export class SearchBarComponent implements OnChanges{
     "Payment Link",
     "Disputes"
   ];
+  @ViewChild(MatAutocompleteTrigger)
+  autocomplete: MatAutocompleteTrigger | undefined;
   filteredOptions: any[] = [];
 
-  ngOnChanges(changes: SimpleChanges) {
 
+  constructor(@Inject(PLATFORM_ID) private platformId: Object) {
+    this.autocomplete = undefined;
+  }
+  
+  ngAfterViewInit() {
+    // Optionally, you might want to add some delay or debounce to avoid too frequent calls
+    if (isPlatformBrowser(this.platformId)) {
+      window.addEventListener('scroll', this.onWindowScroll.bind(this));
+    }
+  }
+
+  @HostListener('window:scroll', ['$event'])
+  onWindowScroll(event: any) {
+    if (this.autocomplete && this.autocomplete.panelOpen) {
+      this.autocomplete.closePanel();
+    }
+  }
+
+  ngOnDestroy() {
+    if (isPlatformBrowser(this.platformId)) {
+      window.removeEventListener('scroll', this.onWindowScroll);
+    }
+  }
+
+  ngOnChanges(changes: SimpleChanges) {
     if(this.searchString.value != "" && this._filter(this.searchString.value || "").length > 0){
       this.inputElement.nativeElement.click();
     }
-    
   }
 
   onSearchChange() {
@@ -59,5 +90,4 @@ export class SearchBarComponent implements OnChanges{
     this.isPanelOpened = filteredItems.length != 0;
     return filteredItems;
   }
-
 }
